Moves editablePanel generator into EditablePanel folder
